vistas sidebar, eliminar-editar clientes y menu lateral en repartidor

+++ resources/views/clientes/index.blade.php
@@ -61,18 +61,22 @@
                         @endif
                       </td>
                       <td>{!! $key->status !!}</td>
-                      <td>
-                         <a href="#" class="btn btn-success btn-sm boton-tabla shadow botonesEditEli" style="border-radius: 5px;" onclick="verCliente('{{$key->id}}','{{$key->qr->codigo}}','{{$key->nombres}}','{{$key->apellidos}}','{{$key->usuario->usuario}}','{{$key->usuario->email}}','{{$key->rut}}')">
-                                      <div class="ti-eye"></div>
-                          </a>
+                      @if(Auth::user()->tipo_usuario == 'Admin' )
+                        <td>
+                           <a href="#" class="btn btn-success btn-sm boton-tabla shadow botonesEditEli" style="border-radius: 5px;" onclick="verCliente('{{$key->id}}','{{$key->qr->codigo}}','{{$key->nombres}}','{{$key->apellidos}}','{{$key->usuario->usuario}}','{{$key->usuario->email}}','{{$key->rut}}')">
+                                        <div class="ti-eye"></div>
+                            </a>
 
-                          <a href="#" class="btn btn-warning btn-sm boton-tabla shadow botonesEditEli" style="border-radius: 5px;" onclick="editarCliente('{{$key->id}}','{{$key->usuario->id}}','{{$key->nombres}}','{{$key->apellidos}}','{{$key->usuario->usuario}}','{{$key->usuario->email}}','{{$key->rut}}','{{$key->status}}')">
-                            <div class="ti-pencil-alt text-white"></div>
-                          </a>
-                          <a href="#" class="btn btn-danger btn-sm boton-tabla shadow botonesEditEli" style="border-radius: 5px;" onclick="eliminarCliente('{{$key->id}}','{{$key->usuario->id}}','{{$key->qr->id}}')">
-                              <div class="ti-trash"></div>
-                          </a>
-                      </td>
+                            <a href="#" class="btn btn-warning btn-sm boton-tabla shadow botonesEditEli" style="border-radius: 5px;" onclick="editarCliente('{{$key->id}}','{{$key->usuario->id}}','{{$key->nombres}}','{{$key->apellidos}}','{{$key->usuario->usuario}}','{{$key->usuario->email}}','{{$key->rut}}','{{$key->status}}')">
+                              <div class="ti-pencil-alt text-white"></div>
+                            </a>
+                            <a href="#" class="btn btn-danger btn-sm boton-tabla shadow botonesEditEli" style="border-radius: 5px;" onclick="eliminarCliente('{{$key->id}}','{{$key->usuario->id}}','{{$key->qr->id}}')">
+                                <div class="ti-trash"></div>
+                            </a>
+                        </td>
+                      @else
+                        <td></td>
+                      @endif
                     </tr>
                   @endforeach
                 </tbody>
